requirements.txt
pycryptodome
mutagen
